keep coding

+++ src/main/java/dearpet/repository/GroomingRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
  * Created by Arianna on 25/11/2017.
  */
 @Repository("groomingRepository")
-public interface GroomingRepository extends JpaRepository<Grooming,Long>{
+public interface GroomingRepository extends JpaRepository<Grooming,Integer>{
 
 }
